fix: read certs and hosts from env through github secrets and vars
